requirements.txt
faiss-cpu
sentence-transformers
langgraph
langchain
python-dotenv
langchain_google_genai
openai